requirements.txt
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==12.0.3
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
whitenoise==5.0.1
